customized bootstrap is added
I took a lot of time to figure out how to customize the variables of bootstrap especially that the project is written in arabic so the RTL (Right to Left) direction needs to be configured properly.
only the primary color is working because just for testing purpose and it works.