item-details: adding some features and editing
Booking breakdown is added (the details of booking)
Date form Group is added to make the user able to choose the dates of renting the product.
the function calculate duration is added to calculate how many months, weeks, days in the dates chosen by the user. so the final price can be calculated based on that.