Chats Tab and No Content Box
Chats tab is added in the profile page, showing all the chats with the user avatar and name and las message.
no content box is added to show it when there is no content available, it is added in the profile page tabs, also added in the results page.

the booking details is passed to the chat to make the owner able to create the booking for the renter to fill it directly in the form if they both agree on it, if not, the owner can change only the fields that need to be changed. (to be done).