Booking managment almost done with minor fixes in other parts
Booking managment page: is added to show all the details of the booking with handling the picking up and returning/finishing the product/service.
the remaining in that part is the review after the booking is complete and done.

settings form: i fixed the problem that prevent the user from updating his data unless he entered the password, so now it works correctly.

rented item card: i fixed it to display real booking data.

also the reviews. number of reviews, and average rating are displayed in all related pages.